Fix: General Scene / Extension
